Rename shouldPipelineRerun.js back to shouldGem2sRerun.js
