+++ src/api.v2/helpers/pipeline/gem2s.js
@@ -16,7 +16,7 @@ const validateRequest = require('../../../utils/schema-validator');
 const getLogger = require('../../../utils/getLogger');
 
 const { qcStepsWithFilterSettings } = require('./pipelineConstruct/qcHelpers');
-const { getGem2sParams, formatSamples } = require('./shouldPipelineRerun');
+const { getGem2sParams, formatSamples } = require('./shouldGem2sRerun');
 const invalidatePlotsForEvent = require('../../../utils/plotConfigInvalidation/invalidatePlotsForEvent');
 const events = require('../../../utils/plotConfigInvalidation/events');
