+++ src/api.v2/helpers/pipeline/seurat.js
@@ -15,7 +15,7 @@ const HookRunner = require('./hooks/HookRunner');
 const validateRequest = require('../../../utils/schema-validator');
 const getLogger = require('../../../utils/getLogger');
 
-const { getGem2sParams, formatSamples } = require('./shouldPipelineRerun');
+const { getGem2sParams, formatSamples } = require('./shouldGem2sRerun');
 
 const logger = getLogger('[SeuratService] - ');
